+++ app/index.tsx
@@ -2,8 +2,15 @@ import { Redirect } from "expo-router";
 import React from "react";
 import { Text } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
+import { useAuth } from "@clerk/clerk-expo";
 
 const Home = () => {
+  const { isSignedIn } = useAuth();
+
+  if (isSignedIn) {
+    return <Redirect href={"/(root)/(tabs)/home"} />;
+  }
+
   return <Redirect href="/(auth)/welcome" />;
 };